# api/docs/swagger.yaml
definitions:
...
  model_booking_service.DoctorTime:
    properties:
      created_at:
        type: string
      department_id:
        type: string
      doctor_date:
        type: string
      doctor_id:
        type: string
      end_time:
        type: string
      id:
        type: integer
      start_time:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
  model_booking_service.DoctorTimesType:
    properties:
      count:
        type: integer
      doctor_times:
        items:
          $ref: '#/definitions/model_booking_service.DoctorTime'
        type: array
    type: object
...
  model_common.ResponseError:
    properties:
      data:
        type: string
      message:
        type: string
      status:
        type: string
    type: object
  model_common.StandardErrorModel:
    properties:
      error:
        $ref: '#/definitions/model_common.ResponseError'
    type: object
...
  model_healthcare_service.DoctorWorkingHoursRes:
    properties:
      created_at:
        type: string
      day_of_week:
        type: string
      doctor_id:
        type: string
      finish_time:
        type: string
      id:
        type: integer
      start_time:
        type: string
      updated_at:
        type: string
    type: object
...
  model_healthcare_service.ListDoctorWorkingHours:
    properties:
      count:
        type: integer
      doctor_working_hours:
        items:
          $ref: '#/definitions/model_healthcare_service.DoctorWorkingHoursRes'
        type: array
    type: object
...
host: localhost:9050
info:
  contact: {}
  title: Dennic Project
  version: "1.7"
paths:
...
  /v1/doctor-time:
    get:
      consumes:
      - application/json
      description: ListDoctorTimes - Api for list doctor time
      parameters:
      - description: search
        enum:
        - status
        in: query
        name: search
        type: string
      - in: query
        name: limit
        type: string
      - in: query
        name: order_by
        type: string
      - in: query
        name: page
        type: string
      - in: query
        name: value
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model_booking_service.DoctorTimesType'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model_common.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model_common.StandardErrorModel'
      summary: ListDoctorTimes
      tags:
      - Doctor Time
  /v1/doctor-time/get:
    get:
      consumes:
      - application/json
      description: GetDoctorTimes - Api for get doctor time
      parameters:
      - description: id
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model_booking_service.DoctorTime'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model_common.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model_common.StandardErrorModel'
      summary: GetDoctorTimes
      tags:
      - Doctor Time
  /v1/doctor-working-hours:
    get:
      consumes:
      - application/json
      description: ListDoctorWorkingHours - Api for list doctor_working_hours
      parameters:
      - in: query
        name: limit
        type: string
      - in: query
        name: order_by
        type: string
      - in: query
        name: page
        type: string
      - in: query
        name: value
        type: string
      - description: search
        enum:
        - day_of_week
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model_healthcare_service.ListDoctorWorkingHours'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model_common.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model_common.StandardErrorModel'
      summary: ListDoctorWorkingHours
      tags:
      - Doctor Working Hours
  /v1/doctor-working-hours/get:
    get:
      consumes:
      - application/json
      description: GetDoctorWorkingHours - Api for get doctor_working_hours
      parameters:
      - description: id
        in: query
        name: id
        required: true
        type: string
      - description: day_of_week
        in: query
        name: day_of_week
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model_healthcare_service.DoctorWorkingHoursRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model_common.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model_common.StandardErrorModel'
      summary: GetDoctorWorkingHours
      tags:
      - Doctor Working Hours
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"